requirements.txt
fastapi
uvicorn
google-generativeai
requests
pillow
python-dotenv
pathlib
Wikipedia-API 
arxiv 
newsapi-python
wikipedia
markdown
pydub
gtts
moviepy
opencv-python
numpy
jinja2
bs4
pymongo[srv]
pdfkit
python-docx
reportlab

langchain-google-genai
chromadb
gitpython
python-multipart
langchain-text-splitters
jinja2
faiss-cpu

jwt
bcrypt
motor
-e .